+++ CHANGES.rst
@@ -14,6 +14,7 @@ Changelog
 - Changed base implementation of MeetingWorkflowConditions.mayDecide to only check if current user has "Review portal content" permission
 - Make the searchlastdecisions meetings search able to display decisions in the future
 - Do not display the 'review_state' columns in contacts dashboard displaying organizations, it is always 'active', we use the 'selected in plonegroup' column information instead
+- Fixed migration of itemInitiators : do not fail if a MeetingUser was deleted and initialize MeetingConfig.orderedContacts and MeetingConfig.orderedItemInitiators correctly
 
 4.1 (2019-08-23)
 ----------------